feat: Add basic admin page
The admin page will include a switch to allow the system to function at required moments or periods throughout the day. It will also include a basic log and  section to add more faces to the system. A basic css was added ass well, with a temporary hardtyped href so that testing could be done easily.